main: added infrastructur to restart a filetransfer
This will be an useful feature when trying to download a large file behind a
unreliable connection. If the connection is lost the filestransfer can be
restartet. This patch only adds the necessary configuration and parsing
code. However, http and ftp methods have to be extended in order to support this
feature.

Signed-off-by: Kurt Kanzenbach <kurt@kmk-computers.de>